remove duplication
